+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Examples (eth-pm): Encrypt Public Key Messages using symmetric encryption. 
 - Guides: Encrypt messages using Waku Message Version 1.
+- Allow passing decryption keys in hex string format.
 
 ### Changed
 - **Breaking**: Moved `startTime` and `endTime` for history queries to a `timeFilter` property as both or neither must be passed; passing only one parameter is not supported.